+++ src/OpenTelemetry.Api/Trace/TracerProviderBuilder.cs
@@ -14,6 +14,7 @@
 // limitations under the License.
 // </copyright>
 using System;
+using System.Diagnostics;
 
 namespace OpenTelemetry.Trace
 {
@@ -45,5 +46,15 @@ namespace OpenTelemetry.Trace
         /// <param name="names">Activity source names.</param>
         /// <returns>Returns <see cref="TracerProviderBuilder"/> for chaining.</returns>
         public abstract TracerProviderBuilder AddSource(params string[] names);
+
+        /// <summary>
+        /// Adds a listener for <see cref="Activity"/> objects created with the given operation name to the <see cref="TracerProviderBuilder"/>.
+        /// </summary>
+        /// <remarks>
+        /// This is provided to capture legacy <see cref="Activity"/> objects created without using the <see cref="System.Diagnostics.ActivitySource"/> API.
+        /// </remarks>
+        /// <param name="operationName">Operation name of the <see cref="Activity"/> objects to capture.</param>
+        /// <returns>Returns <see cref="TracerProviderBuilder"/> for chaining.</returns>
+        public abstract TracerProviderBuilder AddLegacySource(string operationName);
     }
 }