+++ src/OpenTelemetry/Trace/TracerProviderBuilderBase.cs
@@ -79,6 +79,19 @@ namespace OpenTelemetry.Trace
             return this;
         }
 
+        /// <inheritdoc />
+        public override TracerProviderBuilder AddLegacySource(string operationName)
+        {
+            if (string.IsNullOrWhiteSpace(operationName))
+            {
+                throw new ArgumentException($"{nameof(operationName)} contains null or whitespace string.");
+            }
+
+            this.legacyActivityOperationNames[operationName] = true;
+
+            return this;
+        }
+
         /// <summary>
         /// Sets whether the status of <see cref="System.Diagnostics.Activity"/>
         /// should be set to <c>Status.Error</c> when it ended abnormally due to an unhandled exception.
@@ -156,26 +169,6 @@ namespace OpenTelemetry.Trace
             }
 
             this.processors.Add(processor);
-
-            return this;
-        }
-
-        /// <summary>
-        /// Adds a listener for <see cref="Activity"/> objects created with the given operation name to the <see cref="TracerProviderBuilder"/>.
-        /// </summary>
-        /// <remarks>
-        /// This is provided to capture legacy <see cref="Activity"/> objects created without using the <see cref="ActivitySource"/> API.
-        /// </remarks>
-        /// <param name="operationName">Operation name of the <see cref="Activity"/> objects to capture.</param>
-        /// <returns>Returns <see cref="TracerProviderBuilder"/> for chaining.</returns>
-        internal TracerProviderBuilder AddLegacySource(string operationName)
-        {
-            if (string.IsNullOrWhiteSpace(operationName))
-            {
-                throw new ArgumentException($"{nameof(operationName)} contains null or whitespace string.");
-            }
-
-            this.legacyActivityOperationNames[operationName] = true;
 
             return this;
         }