+++ src/OpenTelemetry/Trace/TracerProviderBuilderExtensions.cs
@@ -92,25 +92,6 @@ namespace OpenTelemetry.Trace
         }
 
         /// <summary>
-        /// Adds a listener for <see cref="Activity"/> objects created with the given operation name to the <see cref="TracerProviderBuilder"/>.
-        /// </summary>
-        /// <remarks>
-        /// This is provided to capture legacy <see cref="Activity"/> objects created without using the <see cref="ActivitySource"/> API.
-        /// </remarks>
-        /// <param name="tracerProviderBuilder"><see cref="TracerProviderBuilder"/> instance.</param>
-        /// <param name="operationName">Operation name of the <see cref="Activity"/> objects to capture.</param>
-        /// <returns>Returns <see cref="TracerProviderBuilder"/> for chaining.</returns>
-        public static TracerProviderBuilder AddLegacySource(this TracerProviderBuilder tracerProviderBuilder, string operationName)
-        {
-            if (tracerProviderBuilder is TracerProviderBuilderBase tracerProviderBuilderBase)
-            {
-                tracerProviderBuilderBase.AddLegacySource(operationName);
-            }
-
-            return tracerProviderBuilder;
-        }
-
-        /// <summary>
         /// Run the given actions to initialize the <see cref="TracerProvider"/>.
         /// </summary>
         /// <param name="tracerProviderBuilder"><see cref="TracerProviderBuilder"/>.</param>