+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -9,6 +9,9 @@ please check the latest changes
 
 ## Unreleased
 
+* Adds `AddLegacySource()` to `TracerProviderBuilder`
+  ([#2019](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2019))
+
 ## 1.1.0-beta2
 
 Released 2021-Apr-23